% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_CoxKM_data.R
\docType{data}
\name{CoxKM_data}
\alias{CoxKM_data}
\alias{SNPInfo}
\alias{Z1}
\alias{pheno1}
\title{This is survival data}
\usage{
data(CoxKM_data)
}
\description{
data.frame
}
\details{
\itemize{
  \item pheno1. sex, bmi, time and status for 600 samples
  \item Z1. genotypes for 600 samples for two genes
  \item SNPInfo. SNP name and gene name
}
}

