% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_FbKM_charID.R
\docType{data}
\name{FbKM_charID}
\alias{FbKM_charID}
\alias{fbkm_c_covariates}
\alias{fbkm_c_gene}
\alias{fbkm_c_geneid}
\alias{fbkm_c_weights}
\alias{fbkm_c_y}
\title{This is binary family data with character IDs}
\usage{
data(FbKM_charID)
}
\description{
data.frame
}
\details{
\itemize{
  \item fbkm_c_covariates. age and sex for 100 samples
  \item fbkm_c_gene. genotypes for 100 samples for two genes and each gene has 50 SNPs
  \item fbkm_c_geneid. ordered sample id as in fbkm_c_gene (100 samples)
  \item fbkm_c_weights. weight for each SNP
  \item fbkm_c_y. sample id, sample's father id, sample's mother id and trait
}
}
