% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_FbKM_numID.R
\docType{data}
\name{FbKM_numID}
\alias{FbKM_numID}
\alias{fbkm_n_covariates}
\alias{fbkm_n_gene}
\alias{fbkm_n_geneid}
\alias{fbkm_n_weights}
\alias{fbkm_n_y}
\title{This is binary family data numeric IDs}
\usage{
data(FbKM_numID)
}
\description{
data.frame
}
\details{
\itemize{
  \item fbkm_n_covariates. age and sex for 100 samples
  \item fbkm_n_gene. genotypes for 100 samples for two genes and each gene has 50 SNPs
  \item fbkm_n_geneid. ordered sample id as in fbkm_n_gene (100 samples)
  \item fbkm_n_weights. weight for each SNP
  \item fbkm_n_y. sample id, sample's father id, sample's mother id and trait
}
}
