% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_FKM_charID.R
\docType{data}
\name{FKM_charID}
\alias{FKM_charID}
\alias{fkm_c_covariates}
\alias{fkm_c_gene}
\alias{fkm_c_geneid}
\alias{fkm_c_weights}
\alias{fkm_c_y}
\title{This is continuous family data with character IDs}
\usage{
data(FKM_charID)
}
\description{
data.frame
}
\details{
\itemize{
  \item fkm_c_covariates. age and sex for 100 samples
  \item fkm_c_gene. genotypes for 100 samples for two genes and each gene has 50 SNPs
  \item fkm_c_geneid. ordered sample id as in fkm_c_gene (100 samples)
  \item fkm_c_weights. weight for each SNP
  \item fkm_c_y. sample id, sample's father id, sample's mother id and trait
}
}

