% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_FKM_numID.R
\docType{data}
\name{FKM_numID}
\alias{FKM_numID}
\alias{fkm_n_covariates}
\alias{fkm_n_gene}
\alias{fkm_n_geneid}
\alias{fkm_n_weights}
\alias{fkm_n_y}
\title{This is continuous family data numeric IDs}
\usage{
data(FKM_numID)
}
\description{
data.frame
}
\details{
\itemize{
  \item fkm_n_covariates. age and sex for 100 samples
  \item fkm_n_gene. genotypes for 100 samples for two genes and each gene has 50 SNPs
  \item fkm_n_geneid. ordered sample id as in fkm_n_gene (100 samples)
  \item fkm_n_weights. weight for each SNP
  \item fkm_n_y. sample id, sample's father id, sample's mother id and trait
}
}
