% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_LKM_charID.R
\docType{data}
\name{LKM_charID}
\alias{LKM_charID}
\alias{lkm_c_covariates}
\alias{lkm_c_gene}
\alias{lkm_c_gid}
\alias{lkm_c_weights}
\alias{lkm_c_y}
\title{This is continuous longitudinal data with character IDs}
\usage{
data(LKM_charID)
}
\description{
data.frame
}
\details{
\itemize{
  \item lkm_c_covariates. age and sex for 20 samples, each one has 5 time points
  \item lkm_c_gene. genotypes for 20 samples for two genes and each gene has 50 SNPs
  \item lkm_c_gid. ordered sample id as in lkm_c_gene (20 samples)
  \item lkm_c_weights. weight for each SNP
  \item lkm_c_y. sample id, trait and time points
}
}

