% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LKM_Null_Model.R
\name{LKM_Null_Model}
\alias{LKM_Null_Model}
\title{KM for Quantitative Traits in Longitudinal GWAS Data (fit null model)}
\usage{
LKM_Null_Model(phenotype, time, yid, covariates = NULL)
}
\arguments{
\item{phenotype}{A vector of quantitative trait in the analysis (class: vector). The order should match the vector yid. No missing.}

\item{time}{A vector of time points (class: vector). The order should match the vector yid. No missing.}

\item{yid}{A vector of id (class: vector). Although it doesn't have to be sorted, observations from the same subject have to be connected with each other. The repeated id numbers indicate multiple time points for one subject. Make sure it is not a factor. No missing.}

\item{covariates}{A matrix of covariates (class: data.frame). The order of rows should match the vector yid. Default NULL. No missing.}
}
\value{
output: object as input for LKM
}
\description{
This function (LKM) is used to perform Kernel Machine analysis (Yan et al., 2016) for quantitative traits in GWAS longitudinal data. \cr
# It considers random intercept and random time
}

