% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_LKM_numID.R
\docType{data}
\name{LKM_numID}
\alias{LKM_numID}
\alias{lkm_n_covariates}
\alias{lkm_n_gene}
\alias{lkm_n_gid}
\alias{lkm_n_weights}
\alias{lkm_n_y}
\title{This is continuous longitudinal data with numeric IDs}
\usage{
data(LKM_numID)
}
\description{
data.frame
}
\details{
\itemize{
  \item lkm_n_covariates. age and sex for 20 samples, each one has 5 time points
  \item lkm_n_gene. genotypes for 20 samples for two genes and each gene has 50 SNPs
  \item lkm_n_gid. ordered sample id as in lkm_n_gene (20 samples)
  \item lkm_n_weights. weight for each SNP
  \item lkm_n_y. sample id, trait and time points
}
}

