% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_LKM_numID_int.R
\docType{data}
\name{LKM_numID_int}
\alias{LKM_numID_int}
\alias{lkm_int_n_covariates}
\alias{lkm_int_n_gene}
\alias{lkm_int_n_gid}
\alias{lkm_int_n_weights}
\alias{lkm_int_n_y}
\title{This is continuous longitudinal data with numeric IDs}
\usage{
data(LKM_numID_int)
}
\description{
data.frame
}
\details{
\itemize{
  \item lkm_int_n_covariates. age and sex for 200 samples, each one has 5 time points
  \item lkm_int_n_gene. genotypes for 200 samples for two genes and each gene has 50 SNPs
  \item lkm_int_n_gid. ordered sample id as in lkm_int_n_gene (200 samples)
  \item lkm_int_n_weights. weight for each SNP
  \item lkm_int_n_y. sample id, trait and time points
}
}

