% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_MFKM_charID.R
\docType{data}
\name{MFKM_charID}
\alias{MFKM_charID}
\alias{mfkm_c_covariates}
\alias{mfkm_c_gene}
\alias{mfkm_c_geneid}
\alias{mfkm_c_weights}
\alias{mfkm_c_y}
\title{This is continuous multivariate family data with character IDs}
\usage{
data(MFKM_charID)
}
\description{
data.frame
}
\details{
\itemize{
  \item mfkm_c_covariates. age, sex and pc(trait specific covariate) for 50 samples, each one has two traits
  \item mfkm_c_gene. genotypes for 50 samples for two genes and each gene has 50 SNPs
  \item mfkm_c_geneid. ordered sample id as in mfkm_c_gene (50 samples)
  \item mfkm_c_weights. weight for each SNP
  \item mfkm_c_y. sample id, trait value and trait label
}
}
