% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_MFKM_numID.R
\docType{data}
\name{MFKM_numID}
\alias{MFKM_numID}
\alias{mfkm_n_covariates}
\alias{mfkm_n_gene}
\alias{mfkm_n_geneid}
\alias{mfkm_n_weights}
\alias{mfkm_n_y}
\title{This is continuous multivariate family data with numeric IDs}
\usage{
data(MFKM_numID)
}
\description{
data.frame
}
\details{
\itemize{
  \item mfkm_n_covariates. age, sex and pc(trait specific covariate) for 50 samples, each one has two traits
  \item mfkm_n_gene. genotypes for 50 samples for two genes and each gene has 50 SNPs
  \item mfkm_n_geneid. ordered sample id as in mfkm_n_gene (50 samples)
  \item mfkm_n_weights. weight for each SNP
  \item mfkm_n_y. sample id, trait value and trait label
}
}
