% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_MKM_charID.R
\docType{data}
\name{MKM_charID}
\alias{MKM_charID}
\alias{mkm_c_covariates}
\alias{mkm_c_gene}
\alias{mkm_c_geneid}
\alias{mkm_c_weights}
\alias{mkm_c_ph}
\title{This is continuous multivariate data with character IDs}
\usage{
data(MKM_charID)
}
\description{
data.frame
}
\details{
\itemize{
  \item mkm_c_covariates. age and sex for 50 samples, each one has two traits
  \item mkm_c_gene. genotypes for 50 samples for two genes and each gene has 50 SNPs
  \item mkm_c_geneid. ordered sample id as in mkm_c_gene (50 samples)
  \item mkm_c_weights. weight for each SNP
  \item mkm_c_ph. sample id, trait value and trait label
}
}
