% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_MKM_numID.R
\docType{data}
\name{MKM_numID}
\alias{MKM_numID}
\alias{mkm_n_covariates}
\alias{mkm_n_gene}
\alias{mkm_n_geneid}
\alias{mkm_n_ph}
\alias{mkm_n_weights}
\title{This is continuous multivariate data with numeric IDs}
\usage{
data(MKM_numID)
}
\description{
data.frame
}
\details{
\itemize{
  \item mkm_n_covariates. age and sex for 50 samples, each one has two traits
  \item mkm_n_gene. genotypes for 50 samples for two genes and each gene has 50 SNPs
  \item mkm_n_geneid. ordered sample id as in mkm_n_gene (50 samples)
  \item mkm_n_weights. weight for each SNP
  \item mkm_n_ph. sample id, trait value and trait label
}
}

